third_party/WebKit/LayoutTests/webaudio/audionode-disconnect-expected.txt

Test disconnect() method on AudioNode destination.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Channel #0 contains only the constant 0.
PASS Channel #0 contains only the constant 4.
PASS Channel #0 contains only the constant 2.
PASS Channel #0 contains only the constant 3.
PASS Channel #0 contains only the constant 1.
PASS Channel #1 contains only the constant 2.
PASS Channel #2 contains only the constant 0.
PASS splitter.disconnect(2) threw IndexSizeError: Failed to execute 'disconnect' on 'AudioNode': The output index provided (2) is outside the range [0, 1]..
PASS Disconnecting a connection twice did not throw an exception.
PASS gain1.disconnect(gain2) threw InvalidAccessError: Failed to execute 'disconnect' on 'AudioNode': the given destination is not connected..
PASS gain1.disconnect(gain3) threw InvalidAccessError: Failed to execute 'disconnect' on 'AudioNode': the given destination is not connected..
PASS splitter.disconnect(gain2, 2) threw IndexSizeError: Failed to execute 'disconnect' on 'AudioNode': The output index provided (2) is outside the range [0, 1]..
PASS splitter.disconnect(gain1, 0) threw InvalidAccessError: Failed to execute 'disconnect' on 'AudioNode': output (0) is not connected to the given destination..
PASS splitter.disconnect(gain3, 0, 0) threw InvalidAccessError: Failed to execute 'disconnect' on 'AudioNode': output (0) is not connected to the input (0) of the destination..
PASS splitter.disconnect(merger, 3, 0) threw IndexSizeError: Failed to execute 'disconnect' on 'AudioNode': The output index provided (3) is outside the range [0, 1]..
PASS successfullyParsed is true

TEST COMPLETE
